// src/app/module-sygmak/models/Utilisateur.ts
export class Utilisateur{

 nomEntreprise: string ="";

    nifNom: string ="";
       cmpExpCod: string ="";
    cmpConCod: string ="";
    uuid: string ="";
    nom: string ="";
    phone: string ="";
    email: string ="";
    password: string ="";
    prenom: string ="";
    telephone: string ="";
    roles: string[]=[];
  online?: boolean;   
  rechercheEmail: string = '';
 connectedAt?: string;
  blocked: boolean = false;
   lastConnexion?: string;
    nif: string ="";
      nomenclature: string ="";
      tar_dsc: string ="";
}